# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FlightHub API Documentation",
        "description": "API pour la recherche et la réservation de vols",
        "contact": {
            "name": "FlightHub Support",
            "email": "support@flighthub.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
...
        "/airlines/{airline}": {
            "get": {
                "tags": [
                    "Airlines"
                ],
                "summary": "Récupérer une compagnie aérienne spécifique",
                "description": "Retourne les détails d'une compagnie aérienne par son ID",
                "operationId": "getAirline",
                "parameters": [
                    {
                        "name": "airline",
                        "in": "path",
                        "description": "ID de la compagnie aérienne",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compagnie aérienne récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Airline"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Airline retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compagnie aérienne non trouvée"
                    }
                }
            }
        },
        "/airports": {
            "get": {
                "tags": [
                    "Airports"
                ],
                "summary": "Récupérer la liste des aéroports",
                "description": "Retourne la liste des aéroports avec filtres optionnels",
                "operationId": "getAirports",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filtrer par ville",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city_code",
                        "in": "query",
                        "description": "Filtrer par code de ville",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des aéroports récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Airport"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Airports retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/airports/{airport}": {
            "get": {
                "tags": [
                    "Airports"
                ],
                "summary": "Récupérer un aéroport spécifique",
                "description": "Retourne les détails d'un aéroport par son ID",
                "operationId": "getAirport",
                "parameters": [
                    {
                        "name": "airport",
                        "in": "path",
                        "description": "ID de l'aéroport",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aéroport récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Airport"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Airport retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/airports/search": {
            "get": {
                "tags": [
                    "Airports"
                ],
                "summary": "Rechercher des aéroports",
                "description": "Recherche des aéroports par nom, code IATA ou ville",
                "operationId": "searchAirports",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Terme de recherche",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aéroports trouvés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Airport"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Airports found successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Search query is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Airline": {
                "title": "Compagnie Aérienne",
                "description": "Modèle représentant une compagnie aérienne",
                "properties": {
                    "id": {
                        "description": "ID unique de la compagnie",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom de la compagnie",
                        "type": "string",
                        "example": "Air Canada"
                    },
                    "iata_code": {
                        "description": "Code IATA de la compagnie",
                        "type": "string",
                        "example": "AC"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Airport": {
                "title": "Aéroport",
                "description": "Modèle représentant un aéroport",
                "properties": {
                    "id": {
                        "description": "ID unique de l'aéroport",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom de l'aéroport",
                        "type": "string",
                        "example": "Montréal-Trudeau International Airport"
                    },
                    "iata_code": {
                        "description": "Code IATA de l'aéroport",
                        "type": "string",
                        "example": "YUL"
                    },
                    "city": {
                        "description": "Ville de l'aéroport",
                        "type": "string",
                        "example": "Montréal"
                    },
                    "latitude": {
                        "description": "Latitude de l'aéroport",
                        "type": "number",
                        "format": "float",
                        "example": 45.4706
                    },
                    "longitude": {
                        "description": "Longitude de l'aéroport",
                        "type": "number",
                        "format": "float",
                        "example": -73.7408
                    },
                    "timezone": {
                        "description": "Fuseau horaire de l'aéroport",
                        "type": "string",
                        "example": "America/Montreal"
                    },
                    "city_code": {
                        "description": "Code de la ville",
                        "type": "string",
                        "example": "YMQ"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}